+++ org-code-javabuilder/lang/README.md
@@ -1,0 +1,13 @@
+# Lang
+
+This package is used for code.org-specific implementations of java.lang classes, to be used in user project code.
+
+## Included Classes
+
+- `System` only allows access to a few fields and methods of java.lang.System:
+  - out
+  - in
+  - arrayCopy
+  - currentTimeMillis
+  - lineSeparator
+  - nanoTime
