+++ org-code-javabuilder/settings.gradle
@@ -16,3 +16,5 @@ include('media')
 include('playground')
 include('studentlib')
 include('validation')
+include('lang')
+
